+++ src/test/java/org/folio/search/controller/BrowseAuthorityIT.java
@@ -31,7 +31,7 @@ class BrowseAuthorityIT extends BaseIntegrationTest {
 
   @BeforeAll
   static void prepare() {
-    setUpTenant(31, authorities());
+    setUpTenant(46, authorities());
   }
 
   @AfterAll
@@ -54,10 +54,10 @@ class BrowseAuthorityIT extends BaseIntegrationTest {
   void browseByAuthority_browsingAroundWithAdditionalFilters() {
     var request = get(authorityBrowsePath())
       .param("query", prepareQuery("(headingRef>={value} or headingRef<{value}) "
-                                   + "and isTitleHeadingRef==false "
-                                   + "and tenantId==" + TENANT_ID + " "
-                                   + "and shared==false "
-                                   + "and headingType==(\"Personal Name\")", "\"Ĵämes Röllins\""))
+        + "and isTitleHeadingRef==false "
+        + "and tenantId==" + TENANT_ID + " "
+        + "and shared==false "
+        + "and headingType==(\"Personal Name\")", "\"Ĵämes Röllins\""))
       .param("limit", "7")
       .param("precedingRecordsCount", "2");
     var actual = parseResponse(doGet(request), AuthorityBrowseResult.class);
@@ -78,7 +78,7 @@ class BrowseAuthorityIT extends BaseIntegrationTest {
       .param("precedingRecordsCount", "2");
     var actual = parseResponse(doGet(request), AuthorityBrowseResult.class);
     assertThat(actual).isEqualTo(new AuthorityBrowseResult()
-      .totalRecords(21).prev("Harry Potter").next("Ĵämes Röllins test")
+      .totalRecords(31).prev("Harry Potter").next("Ĵämes Röllins test")
       .items(List.of(
         authorityBrowseItem("Harry Potter", 13, "Uniform Title", AUTHORIZED, 0),
         authorityBrowseItem("Ĵämes Röllins", 2, "Personal Name", REFERENCE, null),
@@ -94,7 +94,7 @@ class BrowseAuthorityIT extends BaseIntegrationTest {
       .param("precedingRecordsCount", "2");
     var actual = parseResponse(doGet(request), AuthorityBrowseResult.class);
     assertThat(actual).isEqualTo(new AuthorityBrowseResult()
-      .totalRecords(21).prev("Fantasy").next("Periodicals")
+      .totalRecords(31).prev("Fantasy").next("Periodicals")
       .items(List.of(
         authorityBrowseItem("Fantasy", 17, "Topical", REFERENCE, null),
         authorityBrowseItem("Harry Potter", 13, "Uniform Title", AUTHORIZED, 0),
@@ -112,9 +112,9 @@ class BrowseAuthorityIT extends BaseIntegrationTest {
       .param("limit", "3");
     var actual = parseResponse(doGet(request), AuthorityBrowseResult.class);
     assertThat(actual).isEqualTo(new AuthorityBrowseResult()
-      .totalRecords(21).prev("War and Peace").next(null)
+      .totalRecords(31).prev("xsftMediumPerfTerm").next(null)
       .items(List.of(
-        authorityBrowseItem("War and Peace", 14, "Uniform Title", REFERENCE, null),
+        authorityBrowseItem("xsftMediumPerfTerm", 35, "Medium of Performance Term", REFERENCE, null),
         authorityBrowseItem("Zappa Frank", 23, "Personal Name", AUTHORIZED, 0).isAnchor(true),
         authorityBrowseItem("Zappa Frank", 24, "Topical", AUTHORIZED, 0).isAnchor(true)
       )));
@@ -129,7 +129,7 @@ class BrowseAuthorityIT extends BaseIntegrationTest {
     var actual = parseResponse(doGet(request), AuthorityBrowseResult.class);
 
     assertThat(actual).isEqualTo(new AuthorityBrowseResult()
-      .totalRecords(21).prev("Disney").next("Ĵämes Röllins")
+      .totalRecords(31).prev("Disney").next("Ĵämes Röllins")
       .items(List.of(
         authorityBrowseItem("Disney", 4, "Corporate Name", AUTHORIZED, 0),
         authorityBrowseItem("Eruption of Vesuvius", 25, "Named Event", AUTHORIZED, 0),
@@ -148,7 +148,7 @@ class BrowseAuthorityIT extends BaseIntegrationTest {
 
     return Stream.of(
       arguments(aroundQuery, "Brian K. Vaughan", 5, new AuthorityBrowseResult()
-        .totalRecords(21).prev("Biomedical Symposium").next("Comic-Con")
+        .totalRecords(31).prev("Biomedical Symposium").next("Comic-Con")
         .items(List.of(
           authorityBrowseItem("Biomedical Symposium", 8, "Conference Name", REFERENCE, null),
           authorityBrowseItem("Blumberg Green Beauty", 5, "Corporate Name", REFERENCE, null),
@@ -157,7 +157,7 @@ class BrowseAuthorityIT extends BaseIntegrationTest {
           authorityBrowseItem("Comic-Con", 7, "Conference Name", AUTHORIZED, 0)))),
 
       arguments(aroundQuery, "harry", 5, new AuthorityBrowseResult()
-        .totalRecords(21).prev("Eruption of Vesuvius").next("Ĵämes Röllins")
+        .totalRecords(31).prev("Eruption of Vesuvius").next("Ĵämes Röllins")
         .items(List.of(
           authorityBrowseItem("Eruption of Vesuvius", 25, "Named Event", AUTHORIZED, 0),
           authorityBrowseItem("Fantasy", 17, "Topical", REFERENCE, null),
@@ -166,21 +166,21 @@ class BrowseAuthorityIT extends BaseIntegrationTest {
           authorityBrowseItem("Ĵämes Röllins", 2, "Personal Name", REFERENCE, null)))),
 
       arguments(aroundQuery, "a", 5, new AuthorityBrowseResult()
-        .totalRecords(21).prev(null).next("Biomedical Symposium")
+        .totalRecords(31).prev(null).next("Biomedical Symposium")
         .items(List.of(
           emptyAuthorityBrowseItem("a"),
           authorityBrowseItem("Asia Pacific", 10, "Geographic Name", AUTHORIZED, 0),
           authorityBrowseItem("Biomedical Symposium", 8, "Conference Name", REFERENCE, null)))),
 
       arguments(aroundQuery, "zz", 5, new AuthorityBrowseResult()
-        .totalRecords(21).prev("Zappa Frank").next(null)
+        .totalRecords(31).prev("Zappa Frank").next(null)
         .items(List.of(
           authorityBrowseItem("Zappa Frank", 24, "Topical", AUTHORIZED, 0),
           authorityBrowseItem("Zappa Frank", 23, "Personal Name", AUTHORIZED, 0),
           emptyAuthorityBrowseItem("zz")))),
 
       arguments(aroundIncludingQuery, "Brian K. Vaughan", 5, new AuthorityBrowseResult()
-        .totalRecords(21).prev("Biomedical Symposium").next("Comic-Con")
+        .totalRecords(31).prev("Biomedical Symposium").next("Comic-Con")
         .items(List.of(
           authorityBrowseItem("Biomedical Symposium", 8, "Conference Name", REFERENCE, null),
           authorityBrowseItem("Blumberg Green Beauty", 5, "Corporate Name", REFERENCE, null),
@@ -189,7 +189,7 @@ class BrowseAuthorityIT extends BaseIntegrationTest {
           authorityBrowseItem("Comic-Con", 7, "Conference Name", AUTHORIZED, 0)))),
 
       arguments(aroundIncludingQuery, "harry", 5, new AuthorityBrowseResult()
-        .totalRecords(21).prev("Eruption of Vesuvius").next("Ĵämes Röllins")
+        .totalRecords(31).prev("Eruption of Vesuvius").next("Ĵämes Röllins")
         .items(List.of(
           authorityBrowseItem("Eruption of Vesuvius", 25, "Named Event", AUTHORIZED, 0),
           authorityBrowseItem("Fantasy", 17, "Topical", REFERENCE, null),
@@ -198,7 +198,7 @@ class BrowseAuthorityIT extends BaseIntegrationTest {
           authorityBrowseItem("Ĵämes Röllins", 2, "Personal Name", REFERENCE, null)))),
 
       arguments(aroundIncludingQuery, "music", 5, new AuthorityBrowseResult()
-        .totalRecords(21).prev("Ĵämes Röllins").next("Novel")
+        .totalRecords(31).prev("Ĵämes Röllins").next("Novel")
         .items(List.of(
           authorityBrowseItem("Ĵämes Röllins", 2, "Personal Name", REFERENCE, null),
           authorityBrowseItem("Knowledge", 29, "General Subdivision", REFERENCE, null),
@@ -207,7 +207,7 @@ class BrowseAuthorityIT extends BaseIntegrationTest {
           authorityBrowseItem("Novel", 19, "Genre", AUTHORIZED, 0)))),
 
       arguments(aroundIncludingQuery, "music", 25, new AuthorityBrowseResult()
-        .totalRecords(21).prev(null).next(null)
+        .totalRecords(31).prev(null).next("xMediumPerfTerm")
         .items(List.of(
           authorityBrowseItem("Asia Pacific", 10, "Geographic Name", AUTHORIZED, 0),
           authorityBrowseItem("Biomedical Symposium", 8, "Conference Name", REFERENCE, null),
@@ -229,12 +229,15 @@ class BrowseAuthorityIT extends BaseIntegrationTest {
           authorityBrowseItem("Revolution", 26, "Named Event", REFERENCE, null),
           authorityBrowseItem("Science", 16, "Topical", AUTHORIZED, 0),
           authorityBrowseItem("War and Peace", 14, "Uniform Title", REFERENCE, null),
-          authorityBrowseItem("Zappa Frank", 23, "Personal Name", AUTHORIZED, 0),
-          authorityBrowseItem("Zappa Frank", 24, "Topical", AUTHORIZED, 0)
+          authorityBrowseItem("xChronSubdivision", 40, "Chronological Subdivision", AUTHORIZED, 0),
+          authorityBrowseItem("xChronTerm", 31, "Chronological Term", AUTHORIZED, 0),
+          authorityBrowseItem("xFormSubdivision", 43, "Form Subdivision", AUTHORIZED, 0),
+          authorityBrowseItem("xGeographicSubdivision", 37, "Geographic Subdivision", AUTHORIZED, 0),
+          authorityBrowseItem("xMediumPerfTerm", 34, "Medium of Performance Term", AUTHORIZED, 0)
         ))),
 
       arguments(aroundIncludingQuery, "music", 11, new AuthorityBrowseResult()
-        .totalRecords(21).prev("Eruption of Vesuvius").next("Revolution")
+        .totalRecords(31).prev("Eruption of Vesuvius").next("Revolution")
         .items(List.of(
           authorityBrowseItem("Eruption of Vesuvius", 25, "Named Event", AUTHORIZED, 0),
           authorityBrowseItem("Fantasy", 17, "Topical", REFERENCE, null),
@@ -249,7 +252,7 @@ class BrowseAuthorityIT extends BaseIntegrationTest {
           authorityBrowseItem("Revolution", 26, "Named Event", REFERENCE, null)))),
 
       arguments(aroundIncludingQuery, "FC", 5, new AuthorityBrowseResult()
-        .totalRecords(21).prev("Eruption of Vesuvius").next("Ĵämes Röllins")
+        .totalRecords(31).prev("Eruption of Vesuvius").next("Ĵämes Röllins")
         .items(List.of(
           authorityBrowseItem("Eruption of Vesuvius", 25, "Named Event", AUTHORIZED, 0),
           authorityBrowseItem("Fantasy", 17, "Topical", REFERENCE, null),
@@ -259,7 +262,7 @@ class BrowseAuthorityIT extends BaseIntegrationTest {
 
       // browsing forward
       arguments(forwardQuery, "Brian K. Vaughan", 5, new AuthorityBrowseResult()
-        .totalRecords(21).prev("Brian K. Vaughan Title").next("Fantasy")
+        .totalRecords(31).prev("Brian K. Vaughan Title").next("Fantasy")
         .items(List.of(
           authorityBrowseItem("Brian K. Vaughan Title", 1, "Personal Name", AUTHORIZED, 0),
           authorityBrowseItem("Comic-Con", 7, "Conference Name", AUTHORIZED, 0),
@@ -268,7 +271,7 @@ class BrowseAuthorityIT extends BaseIntegrationTest {
           authorityBrowseItem("Fantasy", 17, "Topical", REFERENCE, null)))),
 
       arguments(forwardQuery, "biology", 5, new AuthorityBrowseResult()
-        .totalRecords(21).prev("Biomedical Symposium").next("Comic-Con")
+        .totalRecords(31).prev("Biomedical Symposium").next("Comic-Con")
         .items(List.of(
           authorityBrowseItem("Biomedical Symposium", 8, "Conference Name", REFERENCE, null),
           authorityBrowseItem("Blumberg Green Beauty", 5, "Corporate Name", REFERENCE, null),
@@ -278,7 +281,7 @@ class BrowseAuthorityIT extends BaseIntegrationTest {
 
       // checks if collapsing works in forward direction
       arguments(forwardQuery, "F", 5, new AuthorityBrowseResult()
-        .totalRecords(21).prev("Fantasy").next("North America")
+        .totalRecords(31).prev("Fantasy").next("North America")
         .items(List.of(
           authorityBrowseItem("Fantasy", 17, "Topical", REFERENCE, null),
           authorityBrowseItem("Harry Potter", 13, "Uniform Title", AUTHORIZED, 0),
@@ -287,11 +290,11 @@ class BrowseAuthorityIT extends BaseIntegrationTest {
           authorityBrowseItem("North America", 11, "Geographic Name", REFERENCE, null)))),
 
       arguments(forwardQuery, "ZZ", 10, new AuthorityBrowseResult()
-        .totalRecords(21).prev(null).next(null)
+        .totalRecords(31).prev(null).next(null)
         .items(emptyList())),
 
       arguments(forwardIncludingQuery, "Brian K. Vaughan", 5, new AuthorityBrowseResult()
-        .totalRecords(21).prev("Brian K. Vaughan").next("Eruption of Vesuvius")
+        .totalRecords(31).prev("Brian K. Vaughan").next("Eruption of Vesuvius")
         .items(List.of(
           authorityBrowseItem("Brian K. Vaughan", 22, "Personal Name", AUTHORIZED, 0),
           authorityBrowseItem("Brian K. Vaughan Title", 1, "Personal Name", AUTHORIZED, 0),
@@ -300,7 +303,7 @@ class BrowseAuthorityIT extends BaseIntegrationTest {
           authorityBrowseItem("Eruption of Vesuvius", 25, "Named Event", AUTHORIZED, 0)))),
 
       arguments(forwardIncludingQuery, "biology", 5, new AuthorityBrowseResult()
-        .totalRecords(21).prev("Biomedical Symposium").next("Comic-Con")
+        .totalRecords(31).prev("Biomedical Symposium").next("Comic-Con")
         .items(List.of(
           authorityBrowseItem("Biomedical Symposium", 8, "Conference Name", REFERENCE, null),
           authorityBrowseItem("Blumberg Green Beauty", 5, "Corporate Name", REFERENCE, null),
@@ -310,14 +313,14 @@ class BrowseAuthorityIT extends BaseIntegrationTest {
 
       // browsing backward
       arguments(backwardQuery, "Brian K. Vaughan", 5, new AuthorityBrowseResult()
-        .totalRecords(21).prev(null).next("Blumberg Green Beauty")
+        .totalRecords(31).prev(null).next("Blumberg Green Beauty")
         .items(List.of(
           authorityBrowseItem("Asia Pacific", 10, "Geographic Name", AUTHORIZED, 0),
           authorityBrowseItem("Biomedical Symposium", 8, "Conference Name", REFERENCE, null),
           authorityBrowseItem("Blumberg Green Beauty", 5, "Corporate Name", REFERENCE, null)))),
 
       arguments(backwardQuery, "fun", 5, new AuthorityBrowseResult()
-        .totalRecords(21).prev("Brian K. Vaughan Title").next("Fantasy")
+        .totalRecords(31).prev("Brian K. Vaughan Title").next("Fantasy")
         .items(List.of(
           authorityBrowseItem("Brian K. Vaughan Title", 1, "Personal Name", AUTHORIZED, 0),
           authorityBrowseItem("Comic-Con", 7, "Conference Name", AUTHORIZED, 0),
@@ -326,7 +329,7 @@ class BrowseAuthorityIT extends BaseIntegrationTest {
           authorityBrowseItem("Fantasy", 17, "Topical", REFERENCE, null)))),
 
       arguments(backwardQuery, "G", 5, new AuthorityBrowseResult()
-        .totalRecords(21).prev("Brian K. Vaughan Title").next("Fantasy")
+        .totalRecords(31).prev("Brian K. Vaughan Title").next("Fantasy")
         .items(List.of(
           authorityBrowseItem("Brian K. Vaughan Title", 1, "Personal Name", AUTHORIZED, 0),
           authorityBrowseItem("Comic-Con", 7, "Conference Name", AUTHORIZED, 0),
@@ -335,11 +338,11 @@ class BrowseAuthorityIT extends BaseIntegrationTest {
           authorityBrowseItem("Fantasy", 17, "Topical", REFERENCE, null)))),
 
       arguments(backwardQuery, "A", 10, new AuthorityBrowseResult()
-        .totalRecords(21).prev(null).next(null)
+        .totalRecords(31).prev(null).next(null)
         .items(emptyList())),
 
       arguments(backwardIncludingQuery, "Brian K. Vaughan", 5, new AuthorityBrowseResult()
-        .totalRecords(21).prev(null).next("Brian K. Vaughan")
+        .totalRecords(31).prev(null).next("Brian K. Vaughan")
         .items(List.of(
           authorityBrowseItem("Asia Pacific", 10, "Geographic Name", AUTHORIZED, 0),
           authorityBrowseItem("Biomedical Symposium", 8, "Conference Name", REFERENCE, null),
@@ -347,7 +350,7 @@ class BrowseAuthorityIT extends BaseIntegrationTest {
           authorityBrowseItem("Brian K. Vaughan", 22, "Personal Name", AUTHORIZED, 0)))),
 
       arguments(backwardIncludingQuery, "fun", 5, new AuthorityBrowseResult()
-        .totalRecords(21).prev("Brian K. Vaughan Title").next("Fantasy")
+        .totalRecords(31).prev("Brian K. Vaughan Title").next("Fantasy")
         .items(List.of(
           authorityBrowseItem("Brian K. Vaughan Title", 1, "Personal Name", AUTHORIZED, 0),
           authorityBrowseItem("Comic-Con", 7, "Conference Name", AUTHORIZED, 0),
@@ -389,7 +392,22 @@ class BrowseAuthorityIT extends BaseIntegrationTest {
         authority(27).saftNamedEvent(List.of("Stock Market Crash")),
         authority(28).generalSubdivision("Periodicals"),
         authority(29).sftGeneralSubdivision(List.of("Knowledge")),
-        authority(30).saftGeneralSubdivision(List.of("Shrines"))
+        authority(30).saftGeneralSubdivision(List.of("Shrines")),
+        authority(31).chronTerm("xChronTerm"),
+        authority(32).sftChronTerm(List.of("xsftChronTerm")),
+        authority(33).saftChronTerm(List.of("xsaftChronTerm")),
+        authority(34).mediumPerfTerm("xMediumPerfTerm"),
+        authority(35).sftMediumPerfTerm(List.of("xsftMediumPerfTerm")),
+        authority(36).saftMediumPerfTerm(List.of("xsaftMediumPerfTerm")),
+        authority(37).geographicSubdivision("xGeographicSubdivision"),
+        authority(38).sftGeographicSubdivision(List.of("xsftGeographicSubdivision")),
+        authority(39).saftGeographicSubdivision(List.of("xsaftGeographicSubdivision")),
+        authority(40).chronSubdivision("xChronSubdivision"),
+        authority(41).sftChronSubdivision(List.of("xsftChronSubdivision")),
+        authority(42).saftChronSubdivision(List.of("xsaftChronSubdivision")),
+        authority(43).formSubdivision("xFormSubdivision"),
+        authority(44).sftFormSubdivision(List.of("xsftFormSubdivision")),
+        authority(45).saftFormSubdivision(List.of("xsaftFormSubdivision"))
         };
   }