+++ NEWS.md
@@ -23,6 +23,7 @@
 * Call Numbers Browse: Support showing instance title in browse results list ([MSEARCH-948](https://folio-org.atlassian.net/browse/MSEARCH-948))
 * Authority search: Add namedEvent, sftNamedEvent, saftNamedEvent to possible search options ([MSEARCH-958](https://folio-org.atlassian.net/browse/MSEARCH-958))
 * Authority search: Add generalSubdivision, sftGeneralSubdivision, saftGeneralSubdivision to possible search options ([MSEARCH-974](https://folio-org.atlassian.net/browse/MSEARCH-974))
+* Authority search: Add new heading types to possible search options ([MSEARCH-573](https://folio-org.atlassian.net/browse/MSEARCH-573))
 
 ### Bug fixes
 * Remove shelving order calculation for local call-number types